requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
pillow==9.5.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tqdm==4.65.0
wandb==0.15.4
nltk==3.8.1
matplotlib==3.7.1 